\name{CONTundirv8-package}
\alias{CONTundirv8-package}
\alias{CONTundirv8}
\docType{package}
\title{
\packageTitle{CONTundirv8}
}
\description{
\packageDescription{CONTundirv8}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CONTundirv8}
\packageIndices{CONTundirv8}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CONTundirv8}

Maintainer: \packageMaintainer{CONTundirv8}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
